+++ src/Drivers/AbstractDecoder.php
@@ -87,62 +87,6 @@ abstract class AbstractDecoder implements DecoderInterface
     }
 
     /**
-     * Parse data uri
-     */
-    protected function parseDataUri(mixed $input): object
-    {
-        $pattern = "/^data:(?P<mediatype>\w+\/[-+.\w]+)?" .
-            "(?P<parameters>(;[-\w]+=[-\w]+)*)(?P<base64>;base64)?,(?P<data>.*)/";
-
-        $result = preg_match($pattern, (string) $input, $matches);
-
-        return new class ($matches, $result)
-        {
-            /**
-             * @param array<mixed> $matches
-             * @return void
-             */
-            public function __construct(private array $matches, private int|false $result)
-            {
-                //
-            }
-
-            public function isValid(): bool
-            {
-                return (bool) $this->result;
-            }
-
-            public function mediaType(): ?string
-            {
-                if (isset($this->matches['mediatype']) && !empty($this->matches['mediatype'])) {
-                    return $this->matches['mediatype'];
-                }
-
-                return null;
-            }
-
-            public function hasMediaType(): bool
-            {
-                return !empty($this->mediaType());
-            }
-
-            public function isBase64Encoded(): bool
-            {
-                return isset($this->matches['base64']) && $this->matches['base64'] === ';base64';
-            }
-
-            public function data(): ?string
-            {
-                if (isset($this->matches['data']) && !empty($this->matches['data'])) {
-                    return $this->matches['data'];
-                }
-
-                return null;
-            }
-        };
-    }
-
-    /**
      * Parse and return a given file path or throw detailed exception if the path is invalid
      *
      * @throws DecoderException