+++ src/Drivers/Imagick/Decoders/DataUriImageDecoder.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Intervention\Image\Drivers\Imagick\Decoders;
 
+use Intervention\Image\DataUri;
 use Intervention\Image\Exceptions\DecoderException;
 use Intervention\Image\Interfaces\ColorInterface;
 use Intervention\Image\Interfaces\ImageInterface;
@@ -21,16 +22,6 @@ class DataUriImageDecoder extends BinaryImageDecoder
             throw new DecoderException('Data Uri must be of type string.');
         }
 
-        $uri = $this->parseDataUri($input);
-
-        if (!$uri->isValid()) {
-            throw new DecoderException('Input is no valid Data Uri Scheme.');
-        }
-
-        if ($uri->isBase64Encoded()) {
-            return parent::decode(base64_decode($uri->data()));
-        }
-
-        return parent::decode(urldecode($uri->data()));
+        return parent::decode(DataUri::readFromString($input)->data());
     }
 }