+++ tests/Unit/Drivers/AbstractDecoderTest.php
@@ -47,49 +47,6 @@ final class AbstractDecoderTest extends BaseTestCase
         $this->assertEquals('Oliver Vogel', $result->get('IFD0.Artist'));
     }
 
-    public function testParseDataUri(): void
-    {
-        $decoder = new class () extends AbstractDecoder
-        {
-            public function parse(mixed $input): object
-            {
-                return parent::parseDataUri($input);
-            }
-
-            public function decode(mixed $input): ImageInterface|ColorInterface
-            {
-                throw new Exception('');
-            }
-        };
-
-        $result = $decoder->parse(
-            'data:image/gif;foo=bar;base64,R0lGODdhAwADAKIAAAQyrKTy/ByS7AQytLT2/AAAAAAAAAAAACwAAAAAAwADAAADBhgU0gMgAQA7'
-        );
-
-        $this->assertTrue($result->isValid());
-        $this->assertEquals('image/gif', $result->mediaType());
-        $this->assertTrue($result->hasMediaType());
-        $this->assertTrue($result->isBase64Encoded());
-        $this->assertEquals(
-            'R0lGODdhAwADAKIAAAQyrKTy/ByS7AQytLT2/AAAAAAAAAAAACwAAAAAAwADAAADBhgU0gMgAQA7',
-            $result->data()
-        );
-
-        $result = $decoder->parse('data:text/plain;charset=utf-8,test');
-        $this->assertTrue($result->isValid());
-        $this->assertEquals('text/plain', $result->mediaType());
-        $this->assertTrue($result->hasMediaType());
-        $this->assertFalse($result->isBase64Encoded());
-        $this->assertEquals('test', $result->data());
-
-        $result = $decoder->parse('data:;charset=utf-8,');
-        $this->assertTrue($result->isValid());
-        $this->assertNull($result->mediaType());
-        $this->assertFalse($result->hasMediaType());
-        $this->assertFalse($result->isBase64Encoded());
-        $this->assertNull($result->data());
-    }
-
     public function testIsValidBase64(): void
     {
         $decoder = new class () extends AbstractDecoder